fix component announcement
- implement component announce to AnnouncementScreen
- implement component to RapidTestDetail
